Moved start to grace.py
